Added relativePath to processString method
This allows filters to know where the string in the `processString` method comes from.

A useful example would be code linting. 
Right now a linting library can run a validation over the given string but can't report any useful information on where to change something to get rid of the reported information.