Extract out functions nested inside of processAndCacheFile to the prototype.
This provides for more flexibility for subclasses, especially if it does something where it would be helpful to reuse one of these functions.

I came to this via https://github.com/broccolijs/broccoli-filter/compare/master...timmfin:symlink-identical-output, where I wanted to make processFile _not_ write a new file to disk when processString returned the same exact output as the input it was sent. However, I realized that may be too significant of a change to make for existing plugins. So instead, if these previously hidden functions were more available to subclasses, it would be trivial to implement the “symlink indention output” functionality via the subclass (among all sorts of other things).